requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pytest-cov==4.0.0
#pylint==2.15.3
#rust based linter
ruff==0.0.284
boto3==1.24.87
#web
fastapi==0.85.0
uvicorn==0.18.3
#others
Flask==2.0.2
gunicorn
Werkzeug==2.2.2
python-dotenv
requests
openai